refactor: Remove clientId from authorization settings, fix namespaces
