+++ src/Audis.OpenID/Audis.OpenID.Authentication/Configuration/AuthenticationSettings.cs
@@ -1,4 +1,4 @@
-﻿namespace Audis.OpenID.Authentication
+﻿namespace Audis.OpenID.Authentication.Configuration
 {
     public class AuthenticationSettings
     {
