+++ src/Audis.OpenID/Audis.OpenID.Authorization/Configuration/AuthorizationSettings.cs
@@ -6,19 +6,10 @@
 
         public string Audience { get; set; }
 
-        public string ClientId { get; set; }
-
         public void Deconstruct(out string issuer, out string audience)
         {
             issuer = this.Issuer;
             audience = this.Audience;
-        }
-
-        public void Deconstruct(out string issuer, out string audience, out string clientId)
-        {
-            issuer = this.Issuer;
-            audience = this.Audience;
-            clientId = this.ClientId;
         }
     }
 }
